Use pg_isready to check if IT db is up

+++ src/main/scala/fi/vm/sade/valintatulosservice/config/ItPostgres.scala
@@ -9,16 +9,31 @@ class ItPostgres {
   val dataDirName = "valintarekisteri-it-db"
   val dbName = "valintarekisteri"
   val port = 65432
+  val startTimeoutRetries = 30
+  val startTimeoutPollingIntervalMillis = 100
   private val dataDirFile = new File(dataDirName)
   if (!dataDirFile.isDirectory) {
     Files.createDirectory(dataDirFile.toPath)
   }
   val dataDirPath = dataDirFile.getAbsolutePath
   s"initdb -D $dataDirPath".!
-  s"postgres --config_file=postgresql/postgresql.conf -D $dataDirName -p $port".run()
+  private val dbProcess = s"postgres --config_file=postgresql/postgresql.conf -D $dataDirName -p $port".run()
+
+  private def isAcceptingConnections(): Boolean = {
+    s"pg_isready -q -t 1 -h localhost -p $port -d $dbName".! == 0
+  }
 
   def start() {
-    Thread.sleep(5000) // TODO replace for real wait
+    var retries = startTimeoutRetries
+    while (!isAcceptingConnections() && retries > 0) {
+      retries -= 1
+      Thread.sleep(startTimeoutPollingIntervalMillis)
+    }
+    if (!isAcceptingConnections()) {
+      val msg = s"postgres not accepting connections in port $port within $startTimeoutRetries attempts of $startTimeoutPollingIntervalMillis ms intervals"
+      println(msg)
+      throw new RuntimeException(msg)
+    }
     s"dropdb -p $port --if-exists $dbName".!
     s"createdb -p $port $dbName".!
   }
